requirements.txt
# Web app
streamlit>=1.25.0

# LangChain & LLM
langchain>=0.1.0
openai>=1.3.0
tiktoken>=0.5.1

# Embeddings & vecteurs
faiss-cpu>=1.7.4
sentence-transformers>=2.2.2

# Azure
azure-storage-blob>=12.16.0

# Utilitaires
python-dotenv>=1.0.0

langchain-community>=0.0.20

pymupdf>=1.23.0